#2099: Types: Add basic arithmetic operations and use them across codebase

+++ src/vt/group/region/group_range.cc
@@ -134,7 +134,7 @@ Range::Range(Range const& in_other, BoundType in_remove_extent)
   for (auto split = 0; split < num_splits; split++) {
     auto const& child_size = size / num_splits;
     auto hi_bound = split == num_splits - 1 ?
-      hi_ : std::min(static_cast<int>(hi_), cur_lo + child_size*stride_);
+      hi_ : std::min(hi_, cur_lo + child_size*stride_);
     auto r1 = std::make_unique<Range>(cur_lo, BoundType{hi_bound}, stride_);
     apply(std::move(r1));
     cur_lo += BoundType{child_size*stride_};
